+++ modules/Grades/ReportCardComments.php
@@ -199,8 +199,8 @@ if ( ! $_REQUEST['modfunc'] )
 		$sql = "SELECT * FROM REPORT_CARD_COMMENT_CATEGORIES WHERE COURSE_ID='".$_REQUEST['course_id']."' AND SYEAR='".UserSyear()."' AND SCHOOL_ID='".UserSchool()."' ORDER BY SORT_ORDER";
 
 		$functions = array(
-			'TITLE' => 'makeTextInput',
-			'SORT_ORDER' => 'makeTextInput',
+			'TITLE' => '_makeTextInput',
+			'SORT_ORDER' => '_makeTextInput',
 			'COLOR' => '_makeColorInput'
 		);
 
@@ -211,8 +211,8 @@ if ( ! $_REQUEST['modfunc'] )
 		);
 
 		$link['add']['html'] = array(
-			'TITLE' => makeTextInput( '', 'TITLE' ),
-			'SORT_ORDER' => makeTextInput( '', 'SORT_ORDER' ),
+			'TITLE' => _makeTextInput( '', 'TITLE' ),
+			'SORT_ORDER' => _makeTextInput( '', 'SORT_ORDER' ),
 			'COLOR' => _makeColorInput( '', 'COLOR' )
 		);
 
@@ -225,10 +225,10 @@ if ( ! $_REQUEST['modfunc'] )
 	elseif ( $_REQUEST['tab_id']=='-1')
 	{
 		$sql = "SELECT * FROM REPORT_CARD_COMMENTS WHERE SCHOOL_ID='".UserSchool()."' AND SYEAR='".UserSyear()."' AND COURSE_ID IS NULL ORDER BY SORT_ORDER";
-		$functions = array('TITLE' => 'makeTextInput','SORT_ORDER' => 'makeTextInput');
+		$functions = array('TITLE' => '_makeTextInput','SORT_ORDER' => '_makeTextInput');
 		$LO_columns = array('TITLE' => _('Comment'),'SORT_ORDER' => _('Sort Order'));
 
-		$link['add']['html'] = array('TITLE'=>makeTextInput('','TITLE'),'SORT_ORDER'=>makeTextInput('','SORT_ORDER'));
+		$link['add']['html'] = array('TITLE'=>_makeTextInput('','TITLE'),'SORT_ORDER'=>_makeTextInput('','SORT_ORDER'));
 		$link['remove']['link'] = 'Modules.php?modname='.$_REQUEST['modname'].'&modfunc=remove&subject_id='.$_REQUEST['subject_id'].'&course_id='.$_REQUEST['course_id'].'&tab_id=-1';
 		$link['remove']['variables'] = array('id' => 'ID');
 		$link['add']['html']['remove'] = button('add');
@@ -245,7 +245,7 @@ if ( ! $_REQUEST['modfunc'] )
 		foreach ( (array) $codes_RET as $code)
 			$code_select[$code['ID']] = $code['TITLE'];
 
-		$functions = array('TITLE' => 'makeCommentsInput','SCALE_ID' => 'makeCommentsInput','SORT_ORDER' => 'makeCommentsInput');
+		$functions = array('TITLE' => '_makeCommentsInput','SCALE_ID' => '_makeCommentsInput','SORT_ORDER' => '_makeCommentsInput');
 
 		$LO_columns = array('TITLE' => _('Comment'),'SCALE_ID' => _('Code Scale'),'SORT_ORDER' => _('Sort Order'));
 
@@ -260,12 +260,12 @@ if ( ! $_REQUEST['modfunc'] )
 
 			if (User('PROFILE')=='admin' && AllowEdit())
 			{
-				$functions += array('CATEGORY_ID' => 'makeCommentsInput');
+				$functions += array('CATEGORY_ID' => '_makeCommentsInput');
 				$LO_columns += array('CATEGORY_ID' => _('Category'));
 			}
 		}
 
-		$link['add']['html'] = array('TITLE'=>makeCommentsInput('','TITLE'),'SCALE_ID'=>makeCommentsInput('','SCALE_ID'),'SORT_ORDER'=>makeCommentsInput('','SORT_ORDER'));
+		$link['add']['html'] = array('TITLE'=>_makeCommentsInput('','TITLE'),'SCALE_ID'=>_makeCommentsInput('','SCALE_ID'),'SORT_ORDER'=>_makeCommentsInput('','SORT_ORDER'));
 		$link['remove']['link'] = 'Modules.php?modname='.$_REQUEST['modname'].'&modfunc=remove&subject_id='.$_REQUEST['subject_id'].'&course_id='.$_REQUEST['course_id'].'&tab_id='.$_REQUEST['tab_id'];
 		$link['remove']['variables'] = array('id' => 'ID');
 		$link['add']['html']['remove'] = button('add');
@@ -295,13 +295,26 @@ if ( ! $_REQUEST['modfunc'] )
 	echo '</form>';
 }
 
-function makeTextInput($value,$name)
-{	global $THIS_RET;
+function _makeTextInput( $value, $name )
+{
+	global $THIS_RET;
 
-	if ( $THIS_RET['ID'])
+	if ( $THIS_RET['ID'] )
+	{
 		$id = $THIS_RET['ID'];
+	}
 	else
+	{
 		$id = 'new';
+	}
+
+	$extra = '';
+
+	if ( $name === 'TITLE'
+		&& $id !== 'new' )
+	{
+		$extra .= ' required';
+	}
 
 	return TextInput(
 		$value,
@@ -311,15 +324,22 @@ function makeTextInput($value,$name)
 	);
 }
 
-function makeCommentsInput($value,$name)
-{	global $THIS_RET,$category_select,$code_select;
+function _makeCommentsInput( $value, $name )
+{
+	global $THIS_RET,
+		$category_select,
+		$code_select;
 
-	if ( $THIS_RET['ID'])
+	if ( $THIS_RET['ID'] )
+	{
 		$id = $THIS_RET['ID'];
+	}
 	else
+	{
 		$id = 'new';
+	}
 
-	if ( $name=='CATEGORY_ID')
+	if ( $name === 'CATEGORY_ID' )
 	{
 		return SelectInput(
 			$value,
@@ -329,7 +349,7 @@ function makeCommentsInput($value,$name)
 			false
 		);
 	}
-    elseif ( $name=='SCALE_ID')
+    elseif ( $name === 'SCALE_ID' )
     {
 		return SelectInput(
 			$value,
@@ -339,9 +359,18 @@ function makeCommentsInput($value,$name)
 			false
 		);
 	}
-	elseif ( $name=='SORT_ORDER')
+
+	$extra = '';
+
+	if ( $name === 'SORT_ORDER' )
 	{
-		$extra = 'size=5 maxlength=5';
+		$extra .= ' size=5 maxlength=5';
+	}
+
+	if ( $name === 'TITLE'
+		&& $id !== 'new' )
+	{
+		$extra .= ' required';
 	}
 
 	return TextInput(